Rewrite the bumpVersion script to typescript
